- Move tabby to kobold - Neaten up third party dispatching - Fix send button
