+++ srv/adapter/agnaistic.ts
@@ -10,9 +10,10 @@ import { handleHorde } from './horde'
 import { handleKobold } from './kobold'
 import { handleMancer } from './mancer'
 import { handleNovel } from './novel'
-import { getTextgenCompletion, getThirdPartyPayload, handleOoba } from './ooba'
+import { getTextgenCompletion, handleOoba } from './ooba'
 import { handleOAI } from './openai'
 import { handleOpenRouter } from './openrouter'
+import { getThirdPartyPayload } from './payloads'
 import { handlePetals } from './petals'
 import { registerAdapter } from './register'
 import { handleReplicate } from './replicate'
